Add config tool with dotnet-run
To avoid having to align dotnet sdk and  make projects all build together, this is an experiment in running a microservice as a binary so that the aspire sln need have no knowledge of building and running it beyond shelling out to call the dotnet binary.

This worked and we can now fire up the config tool without having any
project references in aspire.