Repo list - set primary remote name to 'origin'
Running a `gitopolis clone` was not resulting in the right branch being
checked out and tracked locally due to the arbitrary nature of which
remote was initially cloned. By marking the primary as origin and using
at least https://github.com/timabell/gitopolis/releases/tag/v1.13.0
we'll now get the "origin" as the tracked copy in the working directory
and any additional remotes added as secondary entries.